Reshuffling functions and joining support offset and roll calculation
