+++ pySC/core/SCfeedbackFirstTurn.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from pySC.core.SCgetBPMreading import SCgetBPMreading
-from pySC.core.SCsetCMs2SetPoints import SCsetCMs2SetPoints
+from pySC.core.SCsetpoints import SCsetCMs2SetPoints
 from pySC.utils.feedback import isNew, isRepro, isTransmit, logLastBPM, goldenDonut
