--- pySC/core/SCinit.py
@@ -1,7 +1,0 @@
-from pySC.classes import SimulatedComissioning
-from at import Lattice
-
-
-def SCinit(RING: Lattice) -> SimulatedComissioning:  # TODO temporarily kept for reference
-    return SimulatedComissioning(RING)
-
