+++ pySC/core/SClocoLib.py
@@ -7,7 +7,7 @@ from pySC.core.SCgetDispersion import SCgetDispersion
 from pySC.core.SCgetModelRM import SCgetModelRM
 from pySC.core.SCgetPinv import SCgetPinv
 from pySC.core.SCgetRespMat import SCgetRespMat
-from pySC.core.SCsetMags2SetPoints import SCsetMags2SetPoints
+from pySC.core.SCsetpoints import SCsetMags2SetPoints
 from pySC.core.SCregisterUpdate import SCupdateMagnets
