+++ pySC/core/SCplotSupport.py
@@ -1,11 +1,12 @@
 import matplotlib.pyplot as plt
 import numpy as np
 from pySC.constants import SUPPORT_TYPES
-from pySC.core.SCgetSupportOffset import SCgetSupportOffset
-from pySC.core.SCgetSupportRoll import SCgetSupportRoll
+from pySC.core.SCgetSupportOffsetRoll import support_offset_and_roll
 from pySC.at_wrapper import atpass, atgetfieldvalues, findspos, findorbit6, findorbit4, atlinopt
 
-def SCplotSupport(SC,fontSize=12,shiftAxes=0.03,xLim=[0, findspos(SC.RING,len(SC.RING)+1)]):
+def SCplotSupport(SC,fontSize=12,shiftAxes=0.03,xLim=None):
+    if xLim is None:
+        xLim=[0, findspos(SC.RING,len(SC.RING)+1)]
     if not hasattr(SC.ORD,'Magnet'):
         raise Exception('Magnets must be registered. Use SCregisterMagnets.')
     elif not hasattr(SC.ORD,'BPM'):
@@ -13,8 +14,8 @@ def SCplotSupport(SC,fontSize=12,shiftAxes=0.03,xLim=[0, findspos(SC.RING,len(SC
     C = findspos(SC.RING,len(SC.RING)+1)
     s = np.linspace(xLim[0],xLim[1],100*(xLim[1]-xLim[0]))
     sPos = findspos(SC.RING,range(1,len(SC.RING)+1))
-    offSupportLine  = SCgetSupportOffset(SC,s)
-    rollSupportLine = SCgetSupportRoll(SC,s)
+    offSupportLine, rollSupportLine  = support_offset_and_roll(SC,s)
+
     i=0
     for ord in SC.ORD.Magnet:
         if sPos[ord-1]>=xLim[0] and sPos[ord-1]<=xLim[1]:
